feat(core): rename the Flow task to Subflow
