+++ core/src/test/resources/flows/valids/task-allow-failure-executable-flow.yml
@@ -3,7 +3,7 @@ namespace: io.kestra.tests
 
 tasks:
   - id: fail
-    type: io.kestra.core.tasks.flows.Flow
+    type: io.kestra.core.tasks.flows.Subflow
     allowFailure: true
     namespace: io.kestra.tests
     flowId: for-each-item-subflow-failed
